Update patch set 5
Patch Set 5:

> Patch Set 4: Code-Review+2
> 
> I'm curious also as to whether this is really needed.  It certainly can't hurt, but I wonder whether we could get away without it.

It is not really necessary for now. It is only needed if we change the default to old format. The cost is it adds some time for make.bash (not much, as building the compiler takes much longer than builder the linker).

I combined it with flag flip, for testing.

Patch-set: 5
Reviewer: Gerrit User 13315 <13315@62eb7196-b449-3ce5-99f1-c037f21e1705>