Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> Thank you for this change Michael. If possible, it would be nice for us to have a test because as per what motivated you to make this change, it can be a subtle regression if we stop panicking on undefined intrinsics. it could be in perhaps another CL even.

Thanks for taking a look Emmanuel. This would be doable if we broke out alias into its own top-level function, but not being familiar with the compiler as a codebase, I'm not sure what the convention is here. This init block has what is effectively its own little DSL for initializing this structure.

What does everyone else think? For now, I'm going to land this change as-is.

Patch-set: 2
Reviewer: Gerrit User 25391 <25391@62eb7196-b449-3ce5-99f1-c037f21e1705>