[dev.link] cmd/link/internal/loader: support cloneToExternal for aux syms
The current implementation of loader.cloneToExternal allows conversion
of text symbols, but doesn't support preserving the aux data on a
symbol if it has aux data (ex: a Go text sym). This patch adds the
machinery for this. When a Go text sym is cloned, it records the
original object index as well as the original local sym index. This
can then be used in the aux data accessor routines.

Change-Id: I515216d480800ef3101a495732a7fb2dbc5ac467