net/http: wait for fetch promise to complete on context cancel
When the request context was cancelled, the Transport.RoundTrip method
could return before the fetch promise resolved. This would cause the
success and failure callback functions to get called after they've
been released, which in turn prints a "call to released function"
error to the console.

Avoid that by waiting for the fetch promise to complete before
returning.

Change-Id: Ied1475e31dcba101b3326521b0cd653dbb345e1d