cmd/compile: combine ssa.html columns with identical contents
Combine columns in ssa.html output if they are identical. There
can now be multiple titles per column which are all clickable to
expand and collapse their column. Ensure phases which are
typically wider (ex. sources + AST) can be scrolled horizontally.
Give collapsed columns some padding for better readability. Some
of the work in this CL was started by Josh Bleecher Snyder and
mailed to me in order to continue to completion.

Updates #37766

Change-Id: I313b0917dc1bafe1eb99d91798ea915e5bcfaae9