cmd/compile: expand blocks before values
Currently the compiler will rewrite blocks and then rewrite values, even if a
block has changed as a result of a rewrite. For most existing architectures
this does not pose an issue, however for riscv64 the multi-control value
branch instructions mean that we can significantly benefit from being able
to rewrite blocks multiple times, before rewriting values.

This changes the rewrite process so that if a block changes, the block
rewriting will be reattempted before continuing on and writing values.

Change-Id: Id9b59f8ee629acbeeaf9d5bc397aff54542b3ced