Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> Hey Josh, I wasn't sure what the process was when it came to picking up your work. I tried checking it out, adding work, and pushing but I could not because you were the original author & ACLs or something along those lines. I am learning as I go with Gerrit and git codereview so please let me know what I can be doing better/more properly. I would not want you to lose credit for your part of the work, so let me know the proper way of committing to your work and I could scrap this CL. In the mean time I worked on a few things, but I do have one question. I wasn't sure what you were referring to in the comment "TODO: interaction with initially folded."

In these circumstances, the easiest thing to do is to delete the Change-Id line, which will cause Gerrit to create a new CL--which is what you've done. As to credit, I'm generally not too worried about it; I care more about the improvements. If desired, you could just add a bit of plain old english text to the commit message.

"interaction with initially folded": some columns are folded initially, and others are not. There's a hard-coded list in there somewhere, by pass name. How should this CL change that? Probably columns should be initially unfolded if any of the passes in a column are initially expanded. This would need implementing. It might require a refactor of that code.

I'm adding Cherry and Keith as reviewers for UI feedback, since I know they also use this UI, and this is a substantive change to it.

Patch-set: 2
Reviewer: Gerrit User 13315 <13315@62eb7196-b449-3ce5-99f1-c037f21e1705>
Reviewer: Gerrit User 5143 <5143@62eb7196-b449-3ce5-99f1-c037f21e1705>