Update patch set 3
Patch Set 3:

> Patch Set 2: Code-Review+2
> 
> (1 comment)
> 
> Code looks ok to me.
> The only way to really review this is to check it in and try it. We can always revert if we don't like the UI.

Hey Keith, sounds good to me. The only piece I think that is missing which Josh had pointed out is making sure to sort out the initially collapsed/expanded set. Now that columns are combined, if there is one with both a phase that should be collapsed initially _and_ a phase that should be expanded initially what is the desired behavior? Also like you said about getting it in, I'd be happy to follow up with another CL that addresses this, or add it to this one before we merge. Let me know!

Patch-set: 3