cmd/cgo,runtime: avoid array clone during cgo call [WIP]
The generated cgo code ended up cloning the array or slice
header. When the array is very large, then this can end up cloning
megabytes of data.

Instead of passing in the array or slice to cgoCheckPointer, it now
passes a pointer to that array or slice.

Benchmark results TODO

Fixes #28454

Change-Id: Id97c7541a2b181e2c9e86bb6cadfc2c92911cae0