Update patch set 4
Patch Set 4:

I've tested this and I noticed there's a lot of space between the
phase names when they are merged, which wastes a lot of space at the
top of the screen when several phases are merged in one column.

Screenshot:

https://gist.github.com/ALTree/438a881bece41b4b08624406dc55e3ae#gistcomment-3234528


Maybe the spacing could be reduced a little?

Patch-set: 4
CC: Gerrit User 5385 <5385@62eb7196-b449-3ce5-99f1-c037f21e1705>