cmd/compile:  allow mid-stack inlining when there is a cycle of recursion
Basic approach to fixing #29737. We still disallow inlining for an
immediately-recursive function, but allow inlining if a function is in a recursion
chain.

If all functions in the recursion chain are simple, then we could inline forever
down the recursion chain (eventually running out of stack on the compiler), so we
add a map to keep track of the functions we have already inlined at a call site.
We stop inlining when we reach a function that we have already inlined in the
recursive chain. Of course, normally the inlining will have stopped earlier,
because of the cost function.

We could also limit the depth of inlining by a simple count (say, limit max
inlining of 10 at any given site). Would that limit other opportunities too much?

TODO: add some more tests. runtime.BenchmarkStackCopyNoCache() is actually already a
good test that triggers the check to stop inlining when we reach the start of the
recursive chain again.

Fixes #29737

Change-Id: I892fa84bb07a947b3125ec8f25ed0e508bf2bdf5