[dev.link] cmd/link: begin converting gentext to loader APIs
Begin the job of converting the linker's Gentext phase over to use
loader APIs. This functionality is currently controlled by a new
(temporary) linker option "-newgentext" (to be removed in a later
patch).

This patch includes most architectures except for s390x and PPC
(these will be added in subsequent patches).

Change-Id: Ic7f55c207dcdbbba657330ef007a72ff7c837416