[dev.link] cmd/link/internal/loader: add loader hooks to record TOC origin
For PPC, when we generate a call stub for a given function F, we need
to keep track of F's version so at to select the same TOC instance for
F's stub as for F itself during relocation processing. In the
sym.Symbol world this association is via the Outer field; this patch
adds support to the loader for recording the association using a new
pair of loader methods.

Change-Id: Ie4a22ec80ea3425b376c5ae320d52fcc2e56da56