Update patch set 2
Patch Set 2:

I agree with Joe Tsai, that the change is subtle and I had to go through the context in order to convince my self, that resetting safeUnquote is the right (or better: save) thing to do.
I am now convinced, because with this safeUnquote can only be used once and for a possible second run the unquoting is properly done for the full byte slice. Therefore I feel confident about this CL.

One question, that is open for me is, if there are other cases, where unquoteBytes is called more than once and we will lose the optimization for the second run. Do you mind to double check this with the benchmarks?
The reason why I think we could loose optimization is, that unquoteBytes is called in method object (https://go-review.googlesource.com/c/go/+/226218/2/src/encoding/json/decode.go#694) right after rescanLiteral (as mentioned in the commit message) but also twice in literalStore (the case related to TextMarshaller: https://go-review.googlesource.com/c/go/+/226218/2/src/encoding/json/decode.go#902 and a second case https://go-review.googlesource.com/c/go/+/226218/2/src/encoding/json/decode.go#946).

Patch-set: 2
CC: Gerrit User 16630 <16630@62eb7196-b449-3ce5-99f1-c037f21e1705>