Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> > Patch Set 3:
> > 
> > > Patch Set 3:
> > > 
> > > Would it be hard for SSA to instead optimize "uint64(x) > c" into "uint32(x) > c" when appropriate? That would seem somewhat more general, and I'd feel a bit more comfortable with that.
> > 
> > I don't think it would be hard. It already happens for x of type uint32. For x of type int32 it looks like it still does the sign extension.
> 
> I see the code in OCONV case to do this, for signed integer, it does OpSignExt*.
> 
> What I try to learn is how we can define when is appropriate to make OpSignExt* -> OpZeroExt*. I did try `if tt.Type.Size() > s.config.PtrSize && tt.Etype == TUINT64` for this specific case, and seems it works.

But yes, I think you're right that TUINT64 is correct, and TUINT is not (if the length is int64 or uint64 on a 32-bit system).
So we shouldn't do this CL and instead figure out what rewrite rules we'd need.

Patch-set: 3
Reviewer: Gerrit User 5200 <5200@62eb7196-b449-3ce5-99f1-c037f21e1705>