Update patch set 1
Patch Set 1: Code-Review-2

Sorry, I'm still not sure we need this. If I understand correctly, I guess the functions that require a large alignment should be special, for which you probably don't want the assembler to insert prologues and do it manually instead, i.e. it is likely that the function will be NOFRAME and NOSPLIT. If you really want a large alignment and also automatic prologue generation, I would guess you'll still want to align the entry of the function, instead of generating the prologue, leaving a big hole and aligning the first user-written instruction. So we can make it that PCALIGN at the entry of function applies to the whole function. Could you explain what is the problem for this approach? Thanks.

Temporary -2 to block submission.

Patch-set: 1
Reviewer: Gerrit User 13315 <13315@62eb7196-b449-3ce5-99f1-c037f21e1705>
Label: Code-Review=-2