[dev.link] cmd/link: add new -dumpsymsat linker option
When debugging linker problems, it is common to instrument the linker
at various spots to dump out information on selected symbols, as a
means of discovering what's happening during the link. This patch
introduces a general-purpose symbol dumping debug utility, intended to
make it possible to grab debugging dumps without having to edit and
rebuild the linker, or keep patches lying around.

The -dumpsymsat flag takes advantage of the recently added linked
phase benchmarking infrastructure, meaning that you can request a
symbol dump at each point the benchmark code marks the start of a
linker phase.

Flag argument is of the form

   ^?stype[,stype]*@phase[,phase]*

where 'stype' is a symbol type (ex: STEXT), the '^' is a tag used to
request relocation dumping, 'phase' is a linker phase (ex: "pclntab",
or "dodata"). The keyword 'all' can be substituted for a symbol type
or a phase name. Examples:

- dump STEXT symbols (with relocs) and SNOPTRBSS symbols at dodata + final:

   ^STEXT,SNOPTRBSS@dodata,final

- dump all symbols at pclntab:

   all@pclntab

- dump SELFROSECT symbols at all phases:

   SELFROSECT@all

The symbol dump format looks roughly like:

    %S236: runtime.assertE2I2<1> t=STEXT sz=163 val=0
      + R0: R_TLS_LE  o=5 a=0 tgt=<nil>
      + R1: R_CALL    o=63 a=0 tgt=runtime.getitab
      + R2: R_CALL    o=154 a=0 tgt=runtime.morestack_noctxt

where 236 is the symbol index and '%' indicates reachability.

Change-Id: I772181948252819dce1919b718ef068640e83b6f