[dev.link] cmd/link: remove file I/O in OutBuf for darwin
To remove OutBuf.Msync, we need to make sure that OutBuf resizes files
appropriately. This change does this for Darwin.

Part of the plan includes eventually adding fallocate calls to reduce
the possibilities of errors on full systems. You'll note this change
doesn't do that yet -- it'll be added in the change that actually
removes Outbuf.Msync.

Eventually EarlyMunmap and all conditional code that depends on it will
be removed.

Change-Id: I4968cc8c260b97d8a4f50b0ef870e4f8299e03c0