Update patch set 8
Patch Set 8:

(26 comments)

Thank you Pj for the change. I've added some suggestions.

Please ensure that we can accept all combinations of:
     REAL SIGN IMAG // IMAG SIGN REAL

I've started a simple format that just accepts for starters
a prototype for ParseComplex at:
    https://play.golang.org/p/sK3o9iBNJ55
and that parses the basic form of:
    REAL(PLAIN_FLOAT) SIGN IMAG(PLAIN_FLOAT) and
its various combinations. Please take a look an see if it sparks
some ideas.

Patch-set: 8
CC: Gerrit User 5137 <5137@62eb7196-b449-3ce5-99f1-c037f21e1705>