Update patch set 2
Patch Set 2:

> Just as an observation: The zoneinfo.zip is created without compression: https://github.com/golang/go/blob/master/lib/time/update.bash#L28
> 
> Just by using '-9' compression the size gets down to 388.183 bytes, or when gzipping the complete file 200.469 bytes.
> 
> When embedding would it make sense to add it compressed, since the stdlib already has the code to decompress?

The standard library has the decompression code, but because of package import requirements the time and time/tzdata packages can't call the existing code.  In order to use it we would have to have a copy of the decompression code in time/tzdata itself.  (Or in the time package, since if we had we could compress $GOROOT/lib/time/zoneinfo.zip.  But then we have to consider: do we want to force every program that uses the time package, which is basically every program, to include the decompression code?)  Even if we keep it only in time/tzdata, it's not a simple drop from 800K to 200K, since the decompression code itself will take up space.

Patch-set: 2
Reviewer: Gerrit User 5206 <5206@62eb7196-b449-3ce5-99f1-c037f21e1705>