runtime:  incorporate Gscan acquire/release into lock ranking order
I added routines that can acquire/release a particular rank without
acquiring/releasing an associated lock. I added lockRankGscan as a rank for
acquiring/releasing the Gscan bit.

castogscanstatus() and casGtoPreemptScan() are acquires of the Gscan bit.
casfrom_Gscanstatus() is a release of the Gscan bit. casgstatus() is like an
acquire and release of the Gscan bit, since it will wait if Gscan bit is currently
set.

Currently, with this setup, we appear to have a cycle between hchan lock and Gscan bit.
(I.e the current lock ranking with Gscan doesn't work.)
Still working on it.

DO NOT REVIEW
DO NOT SUBMIT

Change-Id: I58ce526a74ba856cb42078f7b9901f2832e1d45c