runtime:  incorporate Gscan acquire/release into lock ranking order
I added routines that can acquire/release a particular rank without
acquiring/releasing an associated lock. I added lockRankGscan as a rank for
acquiring/releasing the Gscan bit.

castogscanstatus() and casGtoPreemptScan() are acquires of the Gscan bit.
casfrom_Gscanstatus() is a release of the Gscan bit. casgstatus() is like an
acquire and release of the Gscan bit, since it will wait if Gscan bit is currently
set.

We have a cycle between hchan and Gscan. The acquisition of Gscan and then hchan
only happens in syncadjustsudogs() when the G is suspended, so the main normal
ordering (get hchan, then get Gscan) can't be happening. So, I added a new rank
lockRankHchanLeaf that is used when acquiring hchan locks in syncadjustsudogs.
This ranking is set so no other locks can be acquired except other hchan locks.
hchans.

Change-Id: I58ce526a74ba856cb42078f7b9901f2832e1d45c