Update patch set 3
Patch Set 3:

We've run into a very strange bug inside google3 that has been bisected to this CL.

When generating ppc64le programs, a bunch of the output code appears trashed, overwritten by something. This causes the binaries to crash with "invalid instruction" on startup.

My current theory is that this CL somehow causes a bug in the Go linker, causing the linker to trash its output. Just a theory. My one piece of evidence is that the trashing starts mid-function, and continues to the end of the Go functions, encompassing multiple Go packages. The corruption stops at the first cgo function.

I can't share a failing program at the moment. But just letting you know we might have to revert.

It might warrant another look at this CL to see if you can see any way it might fail.

Patch-set: 3