Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> > Patch Set 3:
> > 
> > when I am next at a computer I will mail a revert of the one line in ssa.go, to get the build green again. (if you need it sooner feel free to go ahead without me.) then I’ll undertake a cleanup of bounded and nonnil.
> 
> Sounds good.

I tried and failed to figure out what exactly was going wrong in the linker with this CL.
GOSSAHASH binary search was failing - I got partway down the binary search, then the problem went away on both child branches :(. Could be it needs the bug to be in multiple functions. Could be I was holding it wrong - applying the GOSSAHASH technique to the linker instead of the test itself is tricky.
In any case, not worth spending more time on it.

Patch-set: 3