Update patch set 1
Patch Set 1:

What problem does this solve?

I often leave off Unlock at the end of tests like this when it should be unnecessary. It has the nice side effect that if my assumptions were wrong, it hangs the test and makes me think what was actually happening in the test.

Patch-set: 1
CC: Gerrit User 5065 <5065@62eb7196-b449-3ce5-99f1-c037f21e1705>