math/big: adjusting spaces in arith_s390x.s file
This CL looks big but it only does formatting changes to arith_s390x.s. The file was formatted
using asmfmt, so there should not be any functional impact. I guess one additional way
to verify that is to run asmfmt locally and check if the output file is the same.

Change-Id: I8b4035ef082a4d0357881869327e25253f2d8be1