Update patch set 3
Patch Set 3: Code-Review+2

(3 comments)

This is a cool feature - here are the screenshot of the goroutines profile.
 
https://i.imgur.com/GvAHZMK.png (pprof graph)
https://i.imgur.com/T2c9suQ.png (the legacy http endpoint)

One concern is when users misuse the labels, the profile size will be big (can be as big as the full goroutine dump). If no one objects, I am thinking to approve and merge this soon.

Patch-set: 3
Reviewer: Gerrit User 5190 <5190@62eb7196-b449-3ce5-99f1-c037f21e1705>
Label: Code-Review=+2